.......... [ZBXNEXT-7921] reverted last commit

+++ ui/include/classes/api/services/CHttpTest.php
@@ -651,11 +651,12 @@ class CHttpTest extends CApiService {
 			$del_httpstepids[] = $db_httpstepitem['httpstepid'];
 		}
 
-		DB::delete('httpstepitem', ['httpstepid' => $del_httpstepids]);
+		if ($del_itemids) {
+			DB::delete('httpstepitem', ['httpstepid' => $del_httpstepids]);
+			DB::delete('httptestitem', ['httptestid' => $del_httptestids]);
 
-		DB::delete('httptestitem', ['httptestid' => $del_httptestids]);
-
-		CItemManager::delete($del_itemids);
+			CItemManager::delete($del_itemids);
+		}
 
 		DB::delete('httptest_field', ['httptestid' => $del_httptestids]);
